+++ mayan/apps/mirroring/handlers.py
@@ -1,0 +1,11 @@
+from __future__ import unicode_literals
+
+from .runtime import cache
+
+
+def handler_document_cache_delete(sender, **kwargs):
+    cache.clear_document(document=kwargs['instance'])
+
+
+def handler_node_cache_delete(sender, **kwargs):
+    cache.clear_node(node=kwargs['instance'])