+++ mayan/apps/mirroring/runtime.py
@@ -1,0 +1,3 @@
+from .caches import IndexFilesystemCache
+
+cache = IndexFilesystemCache()