+++ mayan/apps/mirroring/tests/literals.py
@@ -1,4 +1,7 @@
 from __future__ import absolute_import, unicode_literals
 
+TEST_DOCUMENT_PK = 99
 TEST_NODE_EXPRESSION = 'level_1'
 TEST_NODE_EXPRESSION_MULTILINE = 'first\r\nsecond\r\nthird'
+TEST_NODE_PK = 88
+TEST_PATH = '/test/path'